feat: Add services for the Item entity
Add table names in entities. Add queries in the "ItemRepository" which
select items by specific name, owner or category. Create a method
"updateItemFromDto" in the "ItemMapper" that maps item with its dto
for updating entity attributes in the database. Add methods in the
"ItemService" that update "Item" and return list of items chosen by
name, owner or category.